add: signal関数をまとめた
